Fix: colunas monetarias REAL -> DOUBLE PRECISION (perda de centavos)
REAL e float de 32 bits (~7 digitos significativos): valores na casa dos
milhoes perdiam centavos — o DRE gravava 3.255.700,00 no lugar de
3.255.696,13 (R$ 3,87 de diferenca vs. o ECG). Migracao aplicada no banco
(financeiro, financeiro_previsto, produtos, reposicoes, registros, metas) e
DDLs atualizados para instalacoes novas. Financeiro re-sincronizado: agora
bate exatamente com o ERP.

+++ skills/storage_financeiro.py
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS financeiro (
     loja TEXT NOT NULL,
     grupo TEXT NOT NULL,
     classe TEXT NOT NULL,
-    valor REAL DEFAULT 0,
+    valor DOUBLE PRECISION DEFAULT 0,
     atualizado_em TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE (ano_mes, loja, grupo, classe)
 );
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS financeiro_previsto (
     loja TEXT NOT NULL,
     grupo TEXT NOT NULL,
     classe TEXT NOT NULL,
-    valor REAL DEFAULT 0,
+    valor DOUBLE PRECISION DEFAULT 0,
     atualizado_em TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE (ano_mes, loja, grupo, classe)
 );
